bug fix - missing values should match the type of the variable being read.
git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/trunk_sort_and_time@9618 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 05837a791e22efda35080cec7268cff6cf0fd2f3

+++ observations/utilities/obs_utilities_mod.f90
@@ -419,9 +419,9 @@ subroutine getvar_int(ncid, varname, darray, dmiss)
  integer,            intent(out)  :: darray(:)
  integer,  optional, intent(out)  :: dmiss
 
-integer  :: varid, nfrc
-real(r8) :: fill, miss
-logical  :: found_miss
+integer :: varid, nfrc
+integer :: fill, miss
+logical :: found_miss
 
 ! read the data for the requested array, and get the fill value
 call nc_check( nf90_inq_varid(ncid, varname, varid), &
@@ -699,9 +699,9 @@ subroutine getvar_int_2d(ncid, varname, darray, dmiss)
  integer,            intent(out)  :: darray(:,:)
  integer,  optional, intent(out)  :: dmiss
 
-integer  :: varid, nfrc
-real(r8) :: fill, miss
-logical  :: found_miss
+integer :: varid, nfrc
+integer :: fill, miss
+logical :: found_miss
 
 ! read the data for the requested array, and get the fill value
 call nc_check( nf90_inq_varid(ncid, varname, varid), &
